F01: Add implementation of PartenaireLocalService
