+++ src/main/java/com/carto/sn/service/PartenaireLocalService.java
@@ -1,5 +1,10 @@
 package com.carto.sn.service;
 
-public interface PartenaireLocalService {
+import java.util.List;
 
+import com.carto.sn.entities.PartenaireLocal;
+
+public interface PartenaireLocalService {
+	public List<PartenaireLocal> tousLesPartenairesLocaux();
+	public PartenaireLocal ajoutPartenaireLocalAPartenaire(String nomPartenaireLocal, Long idPartenaire);
 }
